Change setupMenu to setupMenu.js
appのメニュー作成をsetupMenuモジュールに委譲
